+++ nengo/tests/test_connection.py
@@ -394,6 +394,8 @@ def test_slicing(Simulator, nl, plt, seed):
     Ts = [T1, T2, T3]
     ys = [y1, y2, y3]
 
+    weight_solver = nengo.solvers.LstsqL2(weights=True)
+
     with nengo.Network(seed=seed) as m:
         m.config[nengo.Ensemble].neuron_type = nl()
 
@@ -402,10 +404,16 @@ def test_slicing(Simulator, nl, plt, seed):
         nengo.Connection(u, a)
 
         probes = []
+        weight_probes = []
         for sa, sb, T in zip(sas, sbs, Ts):
             b = nengo.Ensemble(N, dimensions=3, radius=1.7)
             nengo.Connection(a[sa], b[sb], transform=T)
             probes.append(nengo.Probe(b, synapse=0.03))
+
+            # also test on weight solver
+            b = nengo.Ensemble(N, dimensions=3, radius=1.7)
+            nengo.Connection(a[sa], b[sb], transform=T, solver=weight_solver)
+            weight_probes.append(nengo.Probe(b, synapse=0.03))
 
     sim = Simulator(m)
     sim.run(0.2)
@@ -417,8 +425,43 @@ def test_slicing(Simulator, nl, plt, seed):
         plt.plot(t, sim.data[p])
 
     atol = 0.01 if nl is nengo.Direct else 0.1
-    for i, [y, p] in enumerate(zip(ys, probes)):
+    for i, [y, p, wp] in enumerate(zip(ys, probes, weight_probes)):
         assert np.allclose(y, sim.data[p][-20:], atol=atol), "Failed %d" % i
+        assert np.allclose(y, sim.data[wp][-20:], atol=atol), "Weights %d" % i
+
+
+def test_neuron_slicing(Simulator, plt, seed, rng):
+    N = 6
+    sa = slice(None, None, 2)
+    sb = slice(None, None, -2)
+
+    x = np.array([-1, -0.25, 1])
+    with nengo.Network(seed=seed) as m:
+        m.config[nengo.Ensemble].neuron_type = nengo.LIFRate()
+
+        u = nengo.Node(output=x)
+        a = nengo.Ensemble(N, dimensions=3, radius=1.7)
+        b = nengo.Ensemble(N, dimensions=3, radius=1.7)
+        nengo.Connection(u, a)
+
+        c = nengo.Connection(a.neurons[sa], b.neurons[sb])
+        c.transform = rng.normal(scale=1e-3, size=(c.size_out, c.size_in))
+
+        ap = nengo.Probe(a.neurons, synapse=0.03)
+        bp = nengo.Probe(b.neurons, synapse=0.03)
+
+    sim = Simulator(m)
+    sim.run(0.2)
+    t = sim.trange()
+
+    x = sim.data[ap]
+    y = np.zeros((len(t), b.n_neurons))
+    y[:, sb] = np.dot(x[:, sa], c.transform.T)
+    y = b.neuron_type.rates(y, sim.data[b].gain, sim.data[b].bias)
+
+    plt.plot(t, y, 'k--')
+    plt.plot(t, sim.data[bp])
+    assert np.allclose(y[-10:], sim.data[bp][-10:], atol=3.0, rtol=0.0)
 
 
 def test_shortfilter(Simulator, nl):
@@ -683,21 +726,11 @@ def test_decoder_probe(Simulator):
     with nengo.Network() as net:
         pre = nengo.Ensemble(10, 10)
         post = nengo.Ensemble(10, 10)
-
         c_ens = nengo.Connection(pre, post)
         c_ens_neurons = nengo.Connection(pre, post.neurons)
-        c_neurons_ens = nengo.Connection(pre.neurons, post)
-        c_neurons = nengo.Connection(pre.neurons, post.neurons)
+        nengo.Probe(c_ens, 'weights')
+        nengo.Probe(c_ens_neurons, 'weights')
 
-        # OK
-        nengo.Probe(c_ens, "decoders")
-        nengo.Probe(c_ens_neurons, "decoders")
-
-        # Not OK
-        with pytest.raises(ValueError):
-            nengo.Probe(c_neurons_ens, "decoders")
-        with pytest.raises(ValueError):
-            nengo.Probe(c_neurons, "decoders")
     assert Simulator(net)
 
 
@@ -712,10 +745,10 @@ def test_transform_probe(Simulator):
         c_neurons_ens = nengo.Connection(pre.neurons, post)
         c_neurons = nengo.Connection(pre.neurons, post.neurons)
 
-        nengo.Probe(c_neurons_ens, "transform")
-        nengo.Probe(c_neurons, "transform")
-        nengo.Probe(c_ens, "transform")
-        nengo.Probe(c_ens_neurons, "transform")
+        nengo.Probe(c_neurons_ens, 'weights')
+        nengo.Probe(c_neurons, 'weights')
+        nengo.Probe(c_ens, 'weights')
+        nengo.Probe(c_ens_neurons, 'weights')
     assert Simulator(net)